moved models into models folder
