+++ account-management/src/main/java/com/aprilhorizon/advice/models/LogResultTuple.java
@@ -1,4 +1,4 @@
-package com.aprilhorizon.advice;
+package com.aprilhorizon.advice.models;
 
 public class LogResultTuple<MethodLog extends LoggedMethod, Object> {
     private MethodLog methodLog;
