+++ account-management/src/main/java/com/aprilhorizon/advice/models/LoggedMethodResponse.java
@@ -1,4 +1,4 @@
-package com.aprilhorizon.advice;
+package com.aprilhorizon.advice.models;
 
 public class LoggedMethodResponse {
     private Object content;
